A rewrite!
The script now takes all options on the command line, rather than using an internal dictionary for configuration.

There is still some configuration needed. If you want to email backups, you have to specify the SMTP server and login information. And, if you really want to, you can store the database information in the script.

The script is now smarter about PostgreSQL, too. If you already have your PostgreSQL Password File setup, you do not have to provide the script a password for the database. Overall, the thing is a bit more robust.