tarsnapper.py: add command-line arguments
Tarsnapper now supports arguments to specify the location of the config
file, delete old archives without creating new ones, and to execute a
single archive rather than all of those that may be defined.

This required a slight refactor of the code. Most notably, the `BACKUP`
variable no longer expects a list of tuples. Instead, it is a simple
dictionary. No changes to old configuration files are required.